rename data to collector so the directory is more action oriented
